2008-10-24 Rodrigo Kumpera <rkumpera@novell.com>
	* Vector16sb.cs: Fix a warning.

	* Vector4f.cs: Add some missing CLSCompliant attributes.

	* Vector8s.cs: Add CompareGreaterThan and a missing
	CLSCompliant attribute.

svn path=/trunk/mcs/; revision=116999

+++ mcs/class/Mono.Simd/Mono.Simd/Vector16sb.cs
@@ -110,7 +110,7 @@ namespace Mono.Simd
 			sbyte *b = &vb.v0;
 			sbyte *c = &res.v0;
 			for (int i = 0; i < 16; ++i)
-				*c++ = (sbyte)(*a++ | *b++);
+				*c++ = (sbyte)((uint)*a++ | (uint)*b++);
 			return res;
 		}
